+++ at_lookup/lib/src/at_lookup_impl.dart
@@ -1,10 +1,11 @@
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-
+import 'dart:typed_data';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:at_commons/src/at_constants.dart' as at_constants;
-import 'package:at_lookup/src/at_lookup.dart';
+import 'package:at_lookup/at_lookup.dart';
 import 'package:at_lookup/src/connection/outbound_connection.dart';
 import 'package:at_lookup/src/connection/outbound_connection_impl.dart';
 import 'package:at_lookup/src/connection/outbound_message_listener.dart';
@@ -18,21 +19,21 @@ class AtLookupImpl implements AtLookUp {
   final logger = AtSignLogger('AtLookup');
 
   /// Listener for reading verb responses from the remote server
-  OutboundMessageListener messageListener;
+  OutboundMessageListener? messageListener;
 
   bool _isPkamAuthenticated = false;
 
   bool _isCramAuthenticated = false;
 
-  OutboundConnection _connection;
+  OutboundConnection? _connection;
 
-  OutboundConnection get connection => _connection;
+  OutboundConnection? get connection => _connection;
 
   var _currentAtSign;
 
   var _rootDomain;
 
-  var _rootPort;
+  late var _rootPort;
 
   var privateKey;
 
@@ -44,8 +45,8 @@ class AtLookupImpl implements AtLookUp {
     String atSign,
     String rootDomain,
     int rootPort, {
-    String privateKey,
-    String cramSecret,
+    String? privateKey,
+    String? cramSecret,
   }) {
     _currentAtSign = atSign;
     _rootDomain = rootDomain;
@@ -54,16 +55,17 @@ class AtLookupImpl implements AtLookUp {
     this.cramSecret = cramSecret;
   }
 
-  static Future<String> findSecondary(
-      String atsign, String rootDomain, int rootPort) async {
-    String response;
-    SecureSocket socket;
+  //#TODO move this static method out
+  static Future<String?> findSecondary(
+      String atsign, String? rootDomain, int rootPort) async {
+    String? response;
+    SecureSocket? socket;
     try {
       AtSignLogger('AtLookup')
           .finer('AtLookup.findSecondary received atsign: $atsign');
       if (atsign.startsWith('@')) atsign = atsign.replaceFirst('@', '');
       var answer = '';
-      String secondary;
+      String? secondary;
       var ans = false;
       var prompt = false;
       var once = true;
@@ -75,7 +77,7 @@ class AtLookupImpl implements AtLookUp {
 
         if (answer.endsWith('@') && prompt == false && once == true) {
           prompt = true;
-          socket.write('$atsign\n');
+          socket!.write('$atsign\n');
           await socket.flush();
           once = false;
         }
@@ -110,7 +112,7 @@ class AtLookupImpl implements AtLookUp {
       throw Exception('AtLookup.findSecondary timed out');
     } on Exception catch (exception) {
       AtSignLogger('AtLookup').severe('AtLookup.findSecondary connection to ' +
-          rootDomain +
+          rootDomain! +
           ' exception: ' +
           exception.toString());
       if (socket != null) {
@@ -128,19 +130,20 @@ class AtLookupImpl implements AtLookUp {
 
   @override
   Future<bool> delete(String key,
-      {String sharedWith, bool isPublic = false}) async {
+      {String? sharedWith, bool isPublic = false}) async {
     var builder = DeleteVerbBuilder()
       ..isPublic = isPublic
       ..sharedWith = sharedWith
       ..atKey = key
       ..sharedBy = _currentAtSign;
     var deleteResult = await executeVerb(builder);
+    // ignore: unnecessary_null_comparison
     return deleteResult != null; //replace with call back
   }
 
   @override
   Future<String> llookup(String key,
-      {String sharedBy, String sharedWith, bool isPublic = false}) async {
+      {String? sharedBy, String? sharedWith, bool isPublic = false}) async {
     var builder;
     if (sharedWith != null) {
       builder = LLookupVerbBuilder()
@@ -172,7 +175,7 @@ class AtLookupImpl implements AtLookUp {
       ..sharedBy = sharedBy
       ..auth = auth
       ..operation = metadata == true ? 'all' : null;
-    if (verifyData == null || verifyData == false) {
+    if (verifyData == false) {
       var lookupResult = await executeVerb(builder);
       lookupResult = VerbUtil.getFormattedValue(lookupResult);
       return lookupResult;
@@ -184,12 +187,12 @@ class AtLookupImpl implements AtLookUp {
         ..sharedBy = sharedBy
         ..auth = false
         ..operation = 'all';
-      var lookupResult = await executeVerb(builder);
+      String? lookupResult = await executeVerb(builder);
       lookupResult = lookupResult.replaceFirst('data:', '');
       var resultJson = json.decode(lookupResult);
       logger.finer(resultJson);
 
-      var publicKeyResult = '';
+      String? publicKeyResult = '';
       if (auth) {
         publicKeyResult = await plookup('publickey', sharedBy);
       } else {
@@ -207,7 +210,7 @@ class AtLookupImpl implements AtLookUp {
       value = VerbUtil.getFormattedValue(value);
       logger.finer('value: $value dataSignature:$dataSignature');
       var isDataValid = publicKey.verifySHA256Signature(
-          utf8.encode(value), base64Decode(dataSignature));
+          utf8.encode(value) as Uint8List, base64Decode(dataSignature));
       logger.finer('atlookup data verify result: $isDataValid');
       return 'data:$value';
     } on Exception catch (e) {
@@ -229,23 +232,19 @@ class AtLookupImpl implements AtLookUp {
 
   @override
   Future<List<String>> scan(
-      {String regex, String sharedBy, bool auth = true}) async {
+      {String? regex, String? sharedBy, bool auth = true}) async {
     var builder = ScanVerbBuilder()
       ..sharedBy = sharedBy
       ..regex = regex
       ..auth = auth;
     var scanResult = await executeVerb(builder);
-    if (scanResult != null) {
-      scanResult = scanResult.replaceFirst('data:', '');
-    }
-    return (scanResult != null && scanResult.isNotEmpty)
-        ? List.from(jsonDecode(scanResult))
-        : [];
+    scanResult = scanResult.replaceFirst('data:', '');
+    return (scanResult.isNotEmpty) ? List.from(jsonDecode(scanResult)) : [];
   }
 
   @override
   Future<bool> update(String key, String value,
-      {String sharedWith, Metadata metadata}) async {
+      {String? sharedWith, Metadata? metadata}) async {
     var builder = UpdateVerbBuilder()
       ..atKey = key
       ..sharedBy = _currentAtSign
@@ -255,12 +254,13 @@ class AtLookupImpl implements AtLookUp {
       builder.ttl = metadata.ttl;
       builder.ttb = metadata.ttb;
       builder.ttr = metadata.ttr;
-      builder.isPublic = metadata.isPublic;
+      builder.isPublic = metadata.isPublic!;
       if (metadata.isHidden) {
         builder.atKey = '_' + key;
       }
     }
     var putResult = await executeVerb(builder);
+    // ignore: unnecessary_null_comparison
     return putResult != null;
   }
 
@@ -279,7 +279,8 @@ class AtLookupImpl implements AtLookUp {
       await createOutBoundConnection(host, port, _currentAtSign);
       //3. listen to server response
       messageListener = OutboundMessageListener(_connection);
-      messageListener.listen();
+      messageListener!.name = 'AtLookUpImpl';
+      messageListener!.listen();
     }
   }
 
@@ -287,7 +288,7 @@ class AtLookupImpl implements AtLookUp {
   /// Optionally [privateKey] is passed for verb builders which require authentication.
   @override
   Future<String> executeVerb(VerbBuilder builder, {sync = false}) async {
-    var verbResult;
+    late var verbResult;
     try {
       if (builder is UpdateVerbBuilder) {
         verbResult = await _update(builder);
@@ -329,11 +330,11 @@ class AtLookupImpl implements AtLookUp {
     return verbResult;
   }
 
-  bool _isError(String verbResult) {
+  bool _isError(String? verbResult) {
     return verbResult != null && verbResult.startsWith('error:');
   }
 
-  Future<String> _update(UpdateVerbBuilder builder) async {
+  Future<String?> _update(UpdateVerbBuilder builder) async {
     var atCommand;
     if (builder.operation == at_constants.UPDATE_META) {
       atCommand = builder.buildCommandForMeta();
@@ -344,63 +345,67 @@ class AtLookupImpl implements AtLookUp {
     return await _process(atCommand, auth: true);
   }
 
-  Future<String> _notify(NotifyVerbBuilder builder) async {
+  Future<String?> _notify(NotifyVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     logger.finer('notify to remote: $atCommand');
     return await _process(atCommand, auth: true);
   }
 
-  Future<String> _scan(ScanVerbBuilder builder) async {
+  Future<String?> _scan(ScanVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(atCommand, auth: builder.auth);
   }
 
-  Future<String> _stats(StatsVerbBuilder builder) async {
+  Future<String?> _stats(StatsVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(atCommand, auth: true);
   }
 
-  Future<String> _config(ConfigVerbBuilder builder) async {
+  Future<String?> _config(ConfigVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(atCommand, auth: true);
   }
 
-  Future<String> _notifyStatus(NotifyStatusVerbBuilder builder) async {
+  Future<String?> _notifyStatus(NotifyStatusVerbBuilder builder) async {
     var command = builder.buildCommand();
     return await _process(command, auth: true);
   }
 
-  Future<String> _notifyList(NotifyListVerbBuilder builder) async {
+  Future<String?> _notifyList(NotifyListVerbBuilder builder) async {
     var command = builder.buildCommand();
     return await _process(command, auth: true);
   }
 
-  Future<String> _notifyAll(NotifyAllVerbBuilder builder) async {
+  Future<String?> _notifyAll(NotifyAllVerbBuilder builder) async {
     var command = builder.buildCommand();
     return await _process(command, auth: true);
   }
 
-  Future<String> executeCommand(String atCommand, {bool auth = false}) async {
+  Future<String?> executeCommand(String atCommand, {bool auth = false}) async {
     return await _process(atCommand, auth: auth);
   }
 
   /// Generates digest using from verb response and [privateKey] and performs a PKAM authentication to
   /// secondary server. This method is executed for all verbs that requires authentication.
-  Future<bool> authenticate(String privateKey) async {
+  Future<bool> authenticate(String? privateKey) async {
     if (privateKey == null) {
       throw UnAuthenticatedException('Private key not passed');
     }
     await _sendCommand('from:$_currentAtSign\n');
-    var fromResponse = await messageListener.read();
+    var fromResponse = await (messageListener!.read());
     logger.finer('from result:$fromResponse');
+    if (fromResponse == null) {
+      return false;
+    }
     fromResponse = fromResponse.trim().replaceAll('data:', '');
     logger.finer('fromResponse $fromResponse');
     var key = RSAPrivateKey.fromString(privateKey);
-    var sha256signature = key.createSHA256Signature(utf8.encode(fromResponse));
+    var sha256signature =
+        key.createSHA256Signature(utf8.encode(fromResponse) as Uint8List);
     var signature = base64Encode(sha256signature);
     logger.finer('Sending command pkam:$signature');
     await _sendCommand('pkam:$signature\n');
-    var pkamResponse = await messageListener.read();
+    var pkamResponse = await messageListener!.read();
     if (pkamResponse == 'data:success') {
       logger.info('auth success');
       _isPkamAuthenticated = true;
@@ -418,14 +423,17 @@ class AtLookupImpl implements AtLookUp {
       throw UnAuthenticatedException('Cram secret not passed');
     }
     await _sendCommand('from:$_currentAtSign\n');
-    var fromResponse = await messageListener.read();
+    var fromResponse = await messageListener!.read();
     logger.info('from result:$fromResponse');
+    if (fromResponse == null) {
+      return false;
+    }
     fromResponse = fromResponse.trim().replaceAll('data:', '');
     var digestInput = '$secret$fromResponse';
     var bytes = utf8.encode(digestInput);
     var digest = sha512.convert(bytes);
     await _sendCommand('cram:$digest\n');
-    var cramResponse = await messageListener.read();
+    var cramResponse = await messageListener!.read();
     if (cramResponse == 'data:success') {
       logger.info('auth success');
       _isCramAuthenticated = true;
@@ -435,22 +443,22 @@ class AtLookupImpl implements AtLookUp {
     return _isCramAuthenticated;
   }
 
-  Future<String> _plookup(PLookupVerbBuilder builder) async {
+  Future<String?> _plookup(PLookupVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(atCommand, auth: true);
   }
 
-  Future<String> _lookup(LookupVerbBuilder builder) async {
+  Future<String?> _lookup(LookupVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(atCommand, auth: builder.auth);
   }
 
-  Future<String> _llookup(LLookupVerbBuilder builder) async {
+  Future<String?> _llookup(LLookupVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(atCommand, auth: true);
   }
 
-  Future<String> _delete(DeleteVerbBuilder builder) async {
+  Future<String?> _delete(DeleteVerbBuilder builder) async {
     var atCommand = builder.buildCommand();
     return await _process(
       atCommand,
@@ -458,8 +466,8 @@ class AtLookupImpl implements AtLookUp {
     );
   }
 
-  Future<String> _process(String command, {bool auth = false}) async {
-    if (auth != null && auth && _isAuthRequired()) {
+  Future<String?> _process(String command, {bool auth = false}) async {
+    if (auth && _isAuthRequired()) {
       if (privateKey != null) {
         await authenticate(privateKey);
       } else if (cramSecret != null) {
@@ -471,7 +479,7 @@ class AtLookupImpl implements AtLookUp {
     }
     try {
       await _sendCommand(command);
-      var result = await messageListener.read();
+      var result = await messageListener!.read();
       return result;
     } on Exception catch (e) {
       logger.severe('Exception in sending to server, ${e.toString()}');
@@ -489,7 +497,7 @@ class AtLookupImpl implements AtLookUp {
       var secureSocket = await SecureSocket.connect(host, int.parse(port));
       _connection = OutboundConnectionImpl(secureSocket);
       if (outboundConnectionTimeout != null) {
-        _connection.setIdleTime(outboundConnectionTimeout);
+        _connection!.setIdleTime(outboundConnectionTimeout);
       }
     } on SocketException {
       throw SecondaryConnectException('unable to connect to secondary');
@@ -498,19 +506,19 @@ class AtLookupImpl implements AtLookUp {
   }
 
   bool isConnectionAvailable() {
-    return _connection != null && !_connection.isInValid();
+    return _connection != null && !_connection!.isInValid();
   }
 
   bool isInValid() {
-    return _connection.isInValid();
+    return _connection!.isInValid();
   }
 
-  Future<void> close() async {
-    await _connection.close();
+  void close() {
+    _connection!.close();
   }
 
   Future<void> _sendCommand(String command) async {
     await createConnection();
-    await _connection.write(command);
+    _connection!.write(command);
   }
 }