+++ at_lookup/lib/src/connection/outbound_connection.dart
@@ -1,11 +1,12 @@
 import 'dart:io';
-import 'package:at_lookup/src/connection/at_connection.dart';
-import 'package:at_lookup/src/connection/base_connection.dart';
-
+import 'package:at_commons/at_commons.dart';
 abstract class OutboundConnection extends BaseConnection {
   OutboundConnection(Socket socket) : super(socket);
-  void setIdleTime(int idleTimeMillis);
+  void setIdleTime(int? idleTimeMillis);
 }
 
 /// Metadata information for [OutboundConnection]
-class OutboundConnectionMetadata extends AtConnectionMetaData {}
+class OutboundConnectionMetadata extends AtConnectionMetaData {
+  String? toAtSign;
+  bool isHandShakeSuccess = false;
+}