+++ at_lookup/lib/src/connection/outbound_connection_impl.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 import 'outbound_connection.dart';
 
 class OutboundConnectionImpl extends OutboundConnection {
-  int outbound_idle_time = 600000; //default timeout 10 minutes
+  int? outbound_idle_time = 600000; //default timeout 10 minutes
 
   OutboundConnectionImpl(Socket socket) : super(socket) {
     metaData = OutboundConnectionMetadata()..created = DateTime.now().toUtc();
@@ -12,11 +12,11 @@ class OutboundConnectionImpl extends OutboundConnection {
     var lastAccessedTime = getMetaData().lastAccessed;
     lastAccessedTime ??= getMetaData().created;
     var currentTime = DateTime.now().toUtc();
-    return currentTime.difference(lastAccessedTime).inMilliseconds;
+    return currentTime.difference(lastAccessedTime!).inMilliseconds;
   }
 
   bool _isIdle() {
-    return _getIdleTimeMillis() > outbound_idle_time;
+    return _getIdleTimeMillis() > outbound_idle_time!;
   }
 
   @override
@@ -25,7 +25,7 @@ class OutboundConnectionImpl extends OutboundConnection {
   }
 
   @override
-  void setIdleTime(int idleTimeMillis) {
+  void setIdleTime(int? idleTimeMillis) {
     outbound_idle_time = idleTimeMillis;
   }
 }